+++ FSCalendar/FSCalendarHeaderView.m
@@ -113,7 +113,7 @@
 
 - (void)setScrollOffset:(CGFloat)scrollOffset animated:(BOOL)animated
 {
-    [self scrollToOffset:scrollOffset animated:NO];
+    [self scrollToOffset:scrollOffset animated:animated];
 }
 
 - (void)scrollToOffset:(CGFloat)scrollOffset animated:(BOOL)animated
@@ -181,8 +181,7 @@
             if ((indexPath.item == 0 || indexPath.item == [self.collectionView numberOfItemsInSection:0] - 1)) {
                 text = nil;
             } else {
-                NSDate *firstPage = [self.calendar.gregorian fs_middleDayOfWeek:self.calendar.minimumDate];
-                NSDate *date = [self.calendar.gregorian dateByAddingUnit:NSCalendarUnitWeekOfYear value:indexPath.item-1 toDate:firstPage options:0];
+                NSDate *date = [self.calendar.calculator pageForSection:indexPath.item - 1];
                 text = [_calendar.formatter stringFromDate:date];
             }
             break;